requirements.txt
aiogram>=3.24.0
telethon>=1.42.0
python-dotenv>=1.0.0
aiofiles>=25.1.0
